Problems with the FinishZone Flag
Tried to add the LoadLevel function for the flag, which was already present via the finishZone.cs. However, it doesn't work, I couldn't get any Debug.Log Info through either. I think there might be something not working with the colliders?
I left everything as it is. Level 1 has the flag right in the beginning now for testing purposes.

All Levels are now added in the BuildSettings.